+++ CHANGES.md
@@ -14,6 +14,7 @@
     - Trailing whitespace characters for line breaks are hidden when using
       `markdown-hide-markup`
     - `fill-paragraph` considers GFM alert syntax [GH-838][]
+    - Add new flag `markdown-wiki-link-retain-case` [GH-839][]
 
 *   Bug fixes:
     - Don't highlight superscript/subscript in math inline/block [GH-802][]
@@ -22,6 +23,7 @@
     - Don't hide backslashes in code blocks when using `markdown-hide-markup`
       and `markdown-fontify-code-blocks-natively` together [GH-766][]
     - Fix `markdown-fontify-whole-heading-line` regression [GH-848][]
+    - Fix using fundamental-mode issue when editting code block [GH-868][]
 
 *   Improvements:
     - Apply url-unescape against URL in an inline link [GH-805][]
@@ -31,6 +33,7 @@
     - Mark `markdown-css-paths` safe as file local variables [GH-834][]
     - Resolve style sheets in `markdown-css-paths` relative to the Markdown file
       (if the path starts with `./` or `../`) [GH-855][] [GH-870][]
+    - Don't insert list item in code block [GH-841][]
 
   [gh-780]: https://github.com/jrblevin/markdown-mode/issues/780
   [gh-802]: https://github.com/jrblevin/markdown-mode/issues/802
@@ -40,9 +43,12 @@
   [gh-827]: https://github.com/jrblevin/markdown-mode/issues/827
   [gh-834]: https://github.com/jrblevin/markdown-mode/issues/834
   [gh-838]: https://github.com/jrblevin/markdown-mode/issues/838
+  [gh-839]: https://github.com/jrblevin/markdown-mode/issues/839
+  [gh-841]: https://github.com/jrblevin/markdown-mode/issues/841
   [gh-845]: https://github.com/jrblevin/markdown-mode/issues/845
   [gh-848]: https://github.com/jrblevin/markdown-mode/issues/848
   [gh-855]: https://github.com/jrblevin/markdown-mode/issues/855
+  [gh-868]: https://github.com/jrblevin/markdown-mode/issues/868
   [gh-870]: https://github.com/jrblevin/markdown-mode/issues/870
 
 # Markdown Mode 2.6